Fix merge conflicts and test failures
- Fix all ruff linting issues (UP038, RUF046, SIM103)
- Resolve mypy type checking errors with proper type ignores
- Fix ZAI connector URL construction and test mocking
- Add missing target_format parameters to translation service calls
- Apply black formatting to all modified files
- Ensure all quality gates pass (black, mypy, ruff, tests)

All ZAI connector tests now passing (5/5)
Code quality tests: 100% green

+++ src/core/app/middleware/exception_middleware.py
@@ -75,4 +75,4 @@ def _build_retry_after_header(reset_at: float | int | None) -> dict[str, str] | 
     if delay_seconds <= 0:
         return {"Retry-After": "0"}
 
-    return {"Retry-After": str(int(math.ceil(delay_seconds)))}
+    return {"Retry-After": str(math.ceil(delay_seconds))}
